Store Pin states as HashMaps
This eliminates the need of passing an `id` field
in the `LocalRequest` everytime. The `respond`
function can handle the request by name using the
HashMap.

This was to done, as unlike the PAD system, the
Local system must hold state for each peripheral.
Using a `OrderedMap` or an `IndexMap` could have
worked, but that would have complicated things
for `serde`.

This also required minor changes in the `setup_pin`
function.