Handle and Respond to LocalRequests
The `sysfs_gpio` crate implements `Send` for the
`Pin` structs, but sharing the entire `LocalState`
(all local gpios in use) across multiple threads
for non determininstic access is not something that
was feasible for the `PAD` and it certainly isn't
worth trying for local peripherals either.

The alternative implementation uses a new task to handle
`LocalRequests` across a `mpsc` channel boundary and sends
responses using a `oneshot` channel.

There is potential to reduce code duplication here by implementing
a trait for this common behaviour.